extract check command methods to manager pkg
